+++ src/aiomysensors/VERSION
