+++ src/aiomysensors/exceptions.py
