+++ src/aiomysensors/model/__init__.py
