+++ src/aiomysensors/model/const.py
