+++ src/aiomysensors/model/message.py
