+++ src/aiomysensors/model/node.py
