+++ src/aiomysensors/model/protocol/protocol_15.py
@@ -2,15 +2,15 @@
 from enum import IntEnum
 
 # pylint: disable=unused-import
+from .protocol_14 import Stream  # noqa: F401
 from .protocol_14 import (  # noqa: F401
     INTERNAL_COMMAND_TYPE,
     STRICT_SYSTEM_COMMAND_TYPES,
     VALID_SYSTEM_COMMAND_TYPES,
     Command,
-    IncomingMessageHandler as IncomingMessageHandler14,
-    OutgoingMessageHandler as OutgoingMessageHandler14,
-    Stream,
 )
+from .protocol_14 import IncomingMessageHandler as IncomingMessageHandler14
+from .protocol_14 import OutgoingMessageHandler as OutgoingMessageHandler14
 
 
 class IncomingMessageHandler(IncomingMessageHandler14):
