+++ src/aiomysensors/transport/serial.py
